Move turn and helper methods to user_turn helper class
